* hm-blender/master: (352 commits)
  glTF: better occlusion map management
  glTF importer: Manage KHR_texture_transform extension
  small try at fixing active_node_material removed from api.
  Another fix to get automated tests pass again in 2.8.
  Blenderkit: remove unused file doing forbidden things.
  glTF exporter: export addon version inside exported gltf/glb files
  Fix lots of various issues failing automated test load/unload/reload.
  Fix indentation, write into temp folder, and expose user prefs
  looptools: fix matrix conversion, temporary fix grease pencil T63677 T61450
  glTF exporter: Keep custom properties when applying modifiers to meshes
  glTF exporter: fix texture image export
  Rigify: change make_property to wrap the official rna_idprop_ui_create.
  archimesh: panel location: Create
  measureit: fix panel warnings, default Tab location: Display
  Pass 3 of 2.8 syntax update...
  glTF importer: Keep user render engine (cycles or Eevee) when importing
  glTF exporter: STEP interpolation if there is only 1 keyframe
  glTF exporter: use extension from filename, not from image name
  matlibvx, fix save library path
  Second pass to update properties for 2.8... To be continued
  ...

+++ io_scene_gltf2/blender/imp/gltf2_blender_image.py
@@ -43,7 +43,7 @@ class BlenderImage():
             if isfile(join(dirname(gltf.filename), pyimage.uri)):
                 return True, join(dirname(gltf.filename), pyimage.uri), basename(join(dirname(gltf.filename), pyimage.uri))
             else:
-                pyimage.gltf.log.error("Missing file (index " + str(img_idx) + "): " + pyimage.uri)
+                gltf.log.error("Missing file (index " + str(img_idx) + "): " + pyimage.uri)
                 return False, None, None
 
         if pyimage.buffer_view is None:
@@ -52,11 +52,13 @@ class BlenderImage():
         return False, None, None
 
     @staticmethod
-    def create(gltf, img_idx):
+    def create(gltf, img_idx, tex_index, tex_transform):
         """Image creation."""
         img = gltf.data.images[img_idx]
 
-        img.blender_image_name = None
+        if img.blender_image_name is not None:
+            # Image is already used somewhere
+            return
 
         if gltf.import_settings['import_pack_images'] is False:
 
@@ -70,10 +72,13 @@ class BlenderImage():
                     if img_.filepath == path:
                         # Already loaded, not needed to reload it
                         img.blender_image_name = img_.name
+                        img_['tex_transform'][str(tex_index)] = tex_transform
                         return
 
                 blender_image = bpy.data.images.load(path)
                 blender_image.name = img_name
+                blender_image['tex_transform'] = {}
+                blender_image['tex_transform'][str(tex_index)] = tex_transform
                 img.blender_image_name = blender_image.name
                 return
 
@@ -83,19 +88,23 @@ class BlenderImage():
             if hasattr(img_, "gltf_index") and img_['gltf_index'] == img_idx:
                 file_creation_needed = False
                 img.blender_image_name = img_.name
+                img_['tex_transform'][tex_index] = tex_transform
                 break
 
         if file_creation_needed is True:
             # Create a temp image, pack, and delete image
             tmp_image = tempfile.NamedTemporaryFile(delete=False)
             img_data, img_name = BinaryData.get_image_data(gltf, img_idx)
-            tmp_image.write(img_data)
-            tmp_image.close()
+            if img_name is not None:
+                tmp_image.write(img_data)
+                tmp_image.close()
 
-            blender_image = bpy.data.images.load(tmp_image.name)
-            blender_image.pack()
-            blender_image.name = img_name
-            img.blender_image_name = blender_image.name
-            blender_image['gltf_index'] = img_idx
-            os.remove(tmp_image.name)
+                blender_image = bpy.data.images.load(tmp_image.name)
+                blender_image.pack()
+                blender_image.name = img_name
+                blender_image['tex_transform'] = {}
+                blender_image['tex_transform'][str(tex_index)] = tex_transform
+                img.blender_image_name = blender_image.name
+                blender_image['gltf_index'] = img_idx
+                os.remove(tmp_image.name)
